test: split thinking provider tests
